memory-agent-chatbot/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
chromadb>=0.4.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
redis>=5.0.0
rich>=13.0.0
